Sensor Fusion - One and Done
- readOdomAngle function added, which finds the heading of the robot using the odometry pod
- getAggregatedHeading function added, which weights the filtered headings of both inertial sensors and the turning odometry pod based on their uncertainties (inertial sensors take priority over the odometry pod)
- type of filteredHeading and filterUncertainty changed to double for neeeded precision

+++ src/kalman.cpp
@@ -130,11 +130,11 @@ KalmanFilter::KalmanFilter(pros::IMU* inertial, pros::Rotation* turnRotational) 
 * but will return -1 if the filter is not active
 * (public)
 */
-int KalmanFilter::getFilteredHeading()
+double KalmanFilter::getFilteredHeading()
     {return this->filteredHeading;}
 
 // (public)
-int KalmanFilter::getFilterUncertainty()
+double KalmanFilter::getFilterUncertainty()
     {return this->filterUncertainty;}
 
 // starts the filter loop if it is not already active (public)
